naver 소셜 로그인 응답값 반환 오류 수정

+++ src/main/java/com/example/fresco/ingredient/controller/dto/request/IngredientFilterRequest.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Setter
 @NoArgsConstructor
 public class IngredientFilterRequest {
-    private List<Long> categoryIds;
+    private List<Short> categoryIds;
     private String sort;
     private Integer page;
     private Integer size;
